+++ packages/nonstop_cli/bricks/flutter_plugin_for_mono_repo/hooks/pre_gen.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 import 'package:mason/mason.dart';
-import 'package:cli_utils/cli_utils.dart';
+import 'package:cli_core/cli_core.dart';
 
 Future<void> run(HookContext context) async {
   // Check if we're in a mono repo using the shared utility